Move commonly useful types to Step/Dir
